scripts/requirements.txt
datetime
mysql-connector-python
python-decouple
requests
xmlrunner